requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
Jinja2==3.1.2
jsql==0.8
libsass==0.22.0
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==2.0.15